Move node.rs and queue.rs to be submodules of list.rs
This move abstracts them to be a sub-module of the inner list.
